Make entity page fully-fledged search
